pyproject.toml
[tool.poetry]
name = "meal-idea-notifier"
version = "0.1.0"
description = ""
authors = ["MNowakowska <mv.nowakowska@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
argparse = "^1.4.0"
pyyaml = "^6.0.1"
pytelegrambotapi = "^4.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
